Safari-611.1.10/LayoutTests/storage/websql/transaction-success-callback-isolated-world-expected.txt

CONSOLE MESSAGE: Web SQL is deprecated. Please use IndexedDB instead.
CONSOLE MESSAGE: Web SQL is deprecated. Please use IndexedDB instead.
ALERT: undefined
ALERT: PASS: document.body.bar visible in a callback created in this world.
This test tests that the transaction success callback is called in the right world.
